requirements.txt
# Dev environment
pip
autopep8

# App
cffi
flask
flask-session
flask-sqlalchemy
flask-login
mysqlclient
bcrypt
pyjwt
mysql
mysql-connector-python

#license plate model
ultralytics
paddleocr
paddlepaddle